backend/requirements.txt
yahoo-finance
newsapi-python
alpha-vantage
fredapi
yfinance
sec_api

langchain
langchain-mistralai 
scikit-learn 
langgraph
bs4
google.generativeai

pydantic 
pydantic_settings 
dotenv 
typing 
setuptools
pandas-ta
langchain_google_genai

fastapi
uvicorn